Add files via upload
Code written in arduino IDE. The robots task was to build a tower from 2.5 x 2.5 [cm] wooden blocks. For this project I used 2 dc motors for driving and 2 servos for controling the "shovel" and "tower" components. The robot faced some problems as the bluetooth module 4.0 is incapable of receiving continous signals which made controlling the robot very limited. Because of lack of time I counld replace the module before competition. The code is simple and easy to understand.